+++ Code/Source/Tools/DioramaEditorModule.cpp
@@ -16,6 +16,7 @@
 #include <Tools/EditorDioramaBulletEmitterComponent.h>
 #include <Tools/EditorDioramaCRTComponent.h>
 #include <Tools/EditorDioramaCamera2DComponent.h>
+#include <Tools/EditorDioramaDepthBodyComponent.h>
 #include <Tools/EditorDioramaHitboxComponent.h>
 #include <Tools/EditorDioramaInputComponent.h>
 #include <Tools/EditorDioramaLightComponent.h>
@@ -57,6 +58,7 @@ namespace Diorama
                     EditorDioramaInputComponent::CreateDescriptor(),
                     EditorDioramaHitboxComponent::CreateDescriptor(),
                     EditorDioramaBulletEmitterComponent::CreateDescriptor(),
+                    EditorDioramaDepthBodyComponent::CreateDescriptor(),
                     EditorDioramaAnimStateMachineComponent::CreateDescriptor(),
                     EditorDioramaAsepriteComponent::CreateDescriptor(),
                     DioramaBuilderComponent::CreateDescriptor(),